pm_career_coach_app/requirements.txt
# Core dependencies for PM Career Coach Streamlit app
streamlit
anthropic
python-dotenv

